+++ projects/hipcub/test/hipcub/test_hipcub_warp_exchange.cpp
@@ -133,11 +133,7 @@ __device__ auto warp_exchange_test(T* d_input, T* d_output)
         thread_data[i] = d_input[threadIdx.x * ItemsPerThread + i];
     }
 
-    using WarpExchangeT                                           = ::hipcub::WarpExchange<T,
-                                                 ItemsPerThread,
-                                                 LogicalWarpSize,
-                                                 1, // ARCH
-                                                 Algorithm>;
+    using WarpExchangeT = ::hipcub::WarpExchange<T, ItemsPerThread, LogicalWarpSize, Algorithm>;
     constexpr unsigned warps_in_block = BlockSize / LogicalWarpSize;
     __shared__ typename WarpExchangeT::TempStorage temp_storage[warps_in_block];
     const unsigned                                 warp_id = threadIdx.x / LogicalWarpSize;